Added utilities to handle basic dialogs
git-svn-id: https://outlook-client.forge.funambol.org/svn/client-sdk/branches/java-sdk@4139 3966be30-755a-0410-815e-dd79f492ac3c

+++ us8294/client/src/main/java/com/funambol/client/ui/DisplayManager.java
@@ -49,7 +49,7 @@ public interface DisplayManager {
      * int value related to the infinitive time to wait before dismissing a
      * screen or a dialog
      */
-    public static final long NO_LIMIT = -1;
+    public static final int NO_LIMIT = -1;
 
     /** Generic dialog id*/
     public static final int GENERIC_DIALOG_ID = 0;
